feat: 식당 통계(평점/저장수) 별점 타입 개선
- 식당 상세 조회 성능 최적화를 위해 `Restaurant` 엔티티에 역정규화된 통계 필드(`totalScraps`, `averageRating`)를 추가
- `UserRestaurantVisitRepository`에 특정 식당의 방문 수와 평균 별점을 집계하는 JPQL 쿼리를 작성
- `RestaurantService`에 `addOrUpdateVisit` 메서드를 구현하여, 사용자가 식당을 저장하거나 별점을 수정할 때마다 실시간으로 통계를 재계산하고 엔티티에 반영
- 별점의 정밀도를 높이기 위해 관련 엔티티 및 DTO의 `rating` 필드 타입을 `Integer`에서 `Double`로 변경

+++ kong-dic/src/main/java/com/kong/kong_dic/domain/user/dto/UserRestaurantVisitResponseDto.java
@@ -4,9 +4,6 @@ import com.kong.kong_dic.domain.restaurant.dto.RestaurantResponseDto;
 import lombok.*;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-// 필요하다면 다른 시간 관련 import 추가 (예: java.time.LocalDateTime)
 
 @Getter @Setter @Builder
 @NoArgsConstructor
@@ -17,6 +14,6 @@ public class UserRestaurantVisitResponseDto {
     private RestaurantResponseDto restaurant;
 
     private LocalDate visitedDate;
-    private Integer rating;
+    private Double rating;
     private String memo;
 }
